Move union-find tests to subpackage
